requirements.txt
# Deep Learning and Reinforcement Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
gymnasium>=0.29.0

# AirSim Integration
airsim>=1.8.1
opencv-python>=4.8.0

# Data Processing and Analysis
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
h5py>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and Utilities
pyyaml>=6.0
jsonschema>=4.17.0
tqdm>=4.65.0
click>=8.1.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# System and File Management
psutil>=5.9.0
pathlib2>=2.3.7

# Development and Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Windows specific dependencies
pywin32>=306; sys_platform == "win32"

# CUDA support (if available)
# Note: CUDA version should match your system
# torch-audio  # Optional for audio processing